Move votes::write to votes::storage::write
