+++ shared/src/ledger/protocol/transactions/votes/storage.rs
@@ -1,0 +1,24 @@
+use borsh::BorshSerialize;
+use eyre::Result;
+
+use super::Tally;
+use crate::ledger::eth_bridge::storage::vote_tallies;
+use crate::ledger::storage::{DBIter, Storage, StorageHasher, DB};
+
+pub fn write<D, H, T>(
+    storage: &mut Storage<D, H>,
+    keys: &vote_tallies::Keys<T>,
+    body: &T,
+    tally: &Tally,
+) -> Result<()>
+where
+    D: 'static + DB + for<'iter> DBIter<'iter> + Sync,
+    H: 'static + StorageHasher + Sync,
+    T: BorshSerialize,
+{
+    storage.write(&keys.body(), &body.try_to_vec()?)?;
+    storage.write(&keys.seen(), &tally.seen.try_to_vec()?)?;
+    storage.write(&keys.seen_by(), &tally.seen_by.try_to_vec()?)?;
+    storage.write(&keys.voting_power(), &tally.voting_power.try_to_vec()?)?;
+    Ok(())
+}
